chat/requirements.txt
openai>=1.0.0
httpx>=0.25.0
pyyaml>=6.0
tomli>=2.0;python_version<"3.11"
rich>=13.0
